requirements.txt
torch==2.7.1
kaleido==0.2.1
planetary-computer==1.0.0
pynvml==12.0.0
python-dotenv==1.1.0
rio-cogeo==5.4.1
stackstac==0.5.1
rioxarray==0.19.0
torchvision==0.22.1
segmentation-models-pytorch==0.5.0
optuna==4.3.0
rasterio==1.4.3
kornia==0.8.1
tqdm==4.67.1
azure-storage-blob==12.25.1